Move two generic task mixins out of backend/core.
There is an external user of ReplTaskMixin, so we leave
a forwarding alias and deprecation message for it.
There is no external usage of MutexTaskMixin, fortunately.
