+++ src/python/pants/task/BUILD
@@ -10,7 +10,9 @@ python_library(
     'src/python/pants/base:exceptions',
     'src/python/pants/base:fingerprint_strategy',
     'src/python/pants/base:worker_pool',
+    'src/python/pants/base:workunit',
     'src/python/pants/cache',
+    'src/python/pants/console:stty_utils',
     'src/python/pants/invalidation',
     'src/python/pants/option',
     'src/python/pants/reporting',
